support fourbyte tracer
