+++ crates/cfxcore/execute-helper/src/observer/mod.rs
@@ -15,6 +15,10 @@ use cfx_vm_tracer_derive::{AsTracer, DrainTrace};
 use std::sync::Arc;
 
 use self::geth_tracer::{GethTracer, TracingInspectorConfig};
+use alloy_rpc_types_trace::geth::{
+    CallConfig, GethDebugBuiltInTracerType, GethDefaultTracingOptions,
+    PreStateConfig,
+};
 
 #[derive(AsTracer, DrainTrace)]
 pub struct Observer {
@@ -50,12 +54,23 @@ impl Observer {
 
     pub fn geth_tracer(
         config: TracingInspectorConfig, tx_gas_limit: u64,
-        machine: Arc<Machine>,
+        machine: Arc<Machine>, tracer_type: Option<GethDebugBuiltInTracerType>,
+        call_config: Option<CallConfig>,
+        prestate_config: Option<PreStateConfig>,
+        opcode_config: Option<GethDefaultTracingOptions>,
     ) -> Self {
         Observer {
             tracer: None,
             gas_man: None,
-            geth_tracer: Some(GethTracer::new(config, tx_gas_limit, machine)),
+            geth_tracer: Some(GethTracer::new(
+                config,
+                tx_gas_limit,
+                machine,
+                tracer_type,
+                call_config,
+                prestate_config,
+                opcode_config,
+            )),
         }
     }
 }
